refactor: refactor GU_email into renamed cookies lib
