+++ src/server/routes/reset.ts
@@ -10,10 +10,9 @@ import { trackMetric } from '@/server/lib/trackMetric';
 import { Metrics } from '@/server/models/Metrics';
 import { PageTitle } from '@/shared/model/PageTitle';
 import { handleAsyncErrors } from '@/server/lib/expressWrappers';
-import { getConfiguration } from '@/server/lib/getConfiguration';
+import { readGUEmailCookie, setGUEmailCookie } from '@/server/lib/cookies';
 
 const router = Router();
-const { baseUri } = getConfiguration();
 
 router.get(Routes.RESET, (req: Request, res: ResponseWithRequestState) => {
   let state = res.locals;
@@ -46,18 +45,7 @@ router.post(
     try {
       await resetPassword(email, req.ip, returnUrl);
 
-      // We set this cookie so that the subsequent email sent page knows which email address was used
-      res.cookie(
-        'GU_email',
-        Buffer.from(JSON.stringify(email)).toString('base64'),
-        // We check if we're running locally here to make testing easier
-        {
-          httpOnly: !baseUri.includes('localhost'),
-          secure: !baseUri.includes('localhost'),
-          signed: !baseUri.includes('localhost'),
-          sameSite: 'strict',
-        },
-      );
+      setGUEmailCookie(res, email);
     } catch (error) {
       logger.error(`${req.method} ${req.originalUrl}  Error`, error);
       const { message, status } = error;
@@ -86,22 +74,7 @@ router.post(
 router.get(Routes.RESET_SENT, (req: Request, res: ResponseWithRequestState) => {
   let state = res.locals;
 
-  // Read the users email from the GU_email cookie that was set when they posted the previous page
-  const emailCookie = baseUri.includes('localhost')
-    ? req.cookies['GU_email']
-    : req.signedCookies['GU_email'];
-
-  let email;
-  try {
-    email = JSON.parse(Buffer.from(emailCookie, 'base64').toString('utf-8'));
-  } catch (error) {
-    email = null;
-    logger.error(
-      `Error parsing cookie with length ${
-        emailCookie ? emailCookie.length : 'undefined'
-      }`,
-    );
-  }
+  const email = readGUEmailCookie(req);
 
   state = deepmerge(state, {
     pageData: {
